requirements.txt
# Core
streamlit
langgraph
langchain-core
langchain-openai
langgraph-checkpoint-sqlite
pydantic

# Gmail OAuth
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2

# Misc
python-dotenv